Converted test inputs to PBN data formats
Updated test infrastructure to pass deal as a single parameter.  Vulnerability is passed to bot also, but there is a mysterious bug that sometimes has the Vulnerability == null but this goes away if you run the test individually.  For now there is code to detect null and pass "None" for the argument.