Game derives from Board class.
Trying to integrate PBN representation with internal representation of classes.  Board class now is a base class for the PBN Game class.  PBN Auction will be next to be updated.  Want to make all of Board, Event name, Board Number, and Auction all classes in Game that get serialized and parsed directly to/from strongly typed classes.  They will only be converted to PBN tags when the game is converted ToString().