IDescribeMultiple constraints plus static descriptions
IDescribeMultipleConstraints receives a list of constraints that are of the same type and must return a List<string> that contains one or more descriptions.  This will be called only once per class TYPE, and any instance may be the one handling the call for all of the constraint instances in a rule.

The ShowsPoints dynamic constraint implements this and will return only one description for points from highest to lowest priority:
   Dummy
   High Card Points
   Starting points

Also, static constraints can now return descriptions.  A few of the static constraints now implement IDescribeConstraint