Better descriptions (#3)
* Min shape shows {min} + for a suit shape

PairMinShape defines the minimum fit with a partnership but there could be more cards held by either partner that have not been shown yet.

Description now has a "+"

* Add global annotations to every remaining call in a call group

It is either add them to every call or change the code that gets annotations for a given call to look at the group.  This seems the most straightforward.

Now Stayman and Jacoby Transfers show up as annotations.  They were there, but the API did not expose them properly.

* IDescribeMultiple constraints plus static descriptions

IDescribeMultipleConstraints receives a list of constraints that are of the same type and must return a List<string> that contains one or more descriptions.  This will be called only once per class TYPE, and any instance may be the one handling the call for all of the constraint instances in a rule.

The ShowsPoints dynamic constraint implements this and will return only one description for points from highest to lowest priority:
   Dummy
   High Card Points
   Starting points

Also, static constraints can now return descriptions.  A few of the static constraints now implement IDescribeConstraint

* Update BridgeBidder/Constraints/PairMinShape.cs

Co-authored-by: Tony Ross <785009+antross@users.noreply.github.com>

* Update BridgeBidder/Constraints/Points.cs

Co-authored-by: Tony Ross <785009+antross@users.noreply.github.com>

---------

Co-authored-by: Tony Ross <785009+antross@users.noreply.github.com>