New test files that were created by the tool
Several of these tests are failiing.  Need to determine how the process of checking in known broken tests shold work.

Perhaps we have a directory with known failing tests.  We check in tests that are known to be correct auctions, but that aren't bid properly yet.  We want the checkin process that verifies the build to pass even when those tests are failing.