New constraints
Added a LogID constraint to help with debugging.  You can put ID("IdString") as a rule constraint and it will show up in the log with that ID in the call description.

Also implemented NewSuit which gives an appropriate description for bidding a new suit.  Removed old NewSuit from OpenBid2.cs