Create static vars for Jump Bids
Where possible we are trying to move to global static variables for constraints that take no parmeters.  New static variables are NonJump, SingleJump and DoubleJump which can be used instead of creating new objects in each rule.

There are more constraints that need to be cleaned up in this manner.  Picking them off one by one.