Massive Clean Up
DynamicConstraint became HandConstriaint.

ShowSuit no longer exists.  Suits are "shown" by the rules that show the shape.  The agreement "show" was redundant and not accurate so it is now gone.

Start of the move to Forcing bids being part of PairAgreements.  Not there yet, but most of the infrastructure is there.

More minor fixes to incorrect bids.

Many more static constraints are now static variables in Bidder.