Suits bid are updated as they are discovered
The FirstToShow method in PairState was searching bids to find the first position that showed a suit.  Now during the update process the hand summary is given to PairState so it keeps it up to date.  No problems with recursion.  Much simpler.

IsReverse now works for any position instead of only opener.